Added login and register pages.
Added a navbar to the user page.

Used Mo's styling for the navbar and login/register pages.

As soon as Angela puts the updated db file in the
repo and y'all import it, we should be able to create accounts.

Next week I'll need to add the ability to become
an admin.